refactor: Rename security.py to securityManager.py
